+++ src/Facades/CryptId.php
@@ -1,12 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tx\CryptId\Facades;
 
 use Illuminate\Support\Facades\Facade;
 
 class CryptId extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return \Tx\CryptId\CryptId::class;
     }