+++ src/TxServiceProvider.php
@@ -1,19 +1,19 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tx\CryptId;
 
 use Illuminate\Support\ServiceProvider;
 
 class TxServiceProvider extends ServiceProvider
 {
-    public function register()
+    public function register(): void
     {
-        $this->app->singleton(CryptId::class, function () {
-            return new CryptId();
-        });
+        $this->app->singleton(CryptId::class, fn (): CryptId => new CryptId());
     }
 
-    public function boot()
+    public function boot(): void
     {
         //
     }